Implement AST and integrate into the grammar
The AST is largely ported from `jsonpath.c` in the Postgres source as of
postgres/postgres@adcdb2c, in development toward the release of Postgres
17 --- although there is a lot yet to do for full stringification.
Whereas Postgres uses a single generic struct for every data type, AST
uses different categories of data types; iterating through the tree will
therefore require type switching as well as some switching on token
types. The forthcoming stringification work will demonstrate (and
perhaps refine) that pattern.

Currently the required method for the Node interface is `String`, which
means any `fmt.Stringer` is allowable, which could cause problems. The
forthcoming stringification work will address this issue, as well.

The output of the `String` method aims to adhere as closely as possible
to the normalized stringification generated by Postgres, including use
of whitespace and escapes --- except where Postgres is a bit more
liberal in what it accepts in identifiers (if you need to use UTF-8 in
identifiers, stringify them).

Use `golang.org/x/tools/cmd/stringer` to generate enum String methods,
both for its efficiency and for the nicety of less boilerplate code to
maintain. Do add tests for them, though (it's amazing the errors I catch
by doing this!)

Adjust the `.golangci.yaml` settings for the `allow` linter to allow
`ast.Node` to be returned, since node accessors that return other node
will not know (or care about) the types of those nodes.

The regular expression constructor adheres as closely as it can to the
spec, especially for the definition of flags, translating them to those
supported by `regexp` and `regexp/syntax`. Like Postgres, the `x` flag
is currently unsupported. Some subtleties will no doubt vary from the
Postgres implementation, as well as [the spec].

Given the reliance on the `regexp` package to validate regular
expressions and flags, the `NewRegex` constructor can return an error,
unlike all the other Node constructors. `NewInteger` and `NewNumeric`
also validate inputs, but panic on error, since data passed from the
parser should always be valid.

Inputs for constructors are otherwise not validated, for the same
reason. If it becomes an issue where we're constructing ASTs that
generate invalid paths, we might have to introduce more validation.

Copy the rest of the SQL/JSON Path grammar from the Postgres source and
replace the C bits with Go bits. The grammar itself is unchanged from
the original except for the addition of the `method` type distinct from
the op_type, because the AST splits them into different structs for
slightly nicer stringification.

Generate the parser and use it in the new `parser.Parse` function, along
with a first test to ensure it works. More tests will be ported from the
Postgres source as stringification improves.

  [the spec]: https://www.w3.org/TR/xpath-functions-3/#flags