Implement and test stringification, fix issues
Add the `writeTo` and `priority` methods to the ast.Node interface, and
implement them to properly generate normalized stringification of
values.

`writeTo` replicates the recursive pattern from PostgreSQL's
`jsonpath.c`, but splits the logic between the different Node
implementations. Note that the use of private methods in the interface
also prevents any other object from outside the package from working as
a node, even if it implements the private methods.

`priority` likewise replicates the `operationPriority` function from
PostgreSQL's `jsonpath.c`, but spread amongst the Node implementations.

In order to better support numeric normalization, replace `rawString`
with `numberNode` and have it store both the literal representation of
the number as scanned and a normalized representation as parsed and
formatted. Use `encoding/json` to format numeric values, to better
approximate what PostgreSQL does, the only difference being that
PostgreSQL preserves trailing zeros in decimal values while Go does not
(I even tried `big.Rat` and `github.com/shopspring/decimal`, and neither
preserves precisions with trailing zeros). In other words, Postgres
formats `0.00` as `0.00` while Go formats it as `0`. Comments in the
tests note this difference.

Modify NewAccessor to merge and return an AccessorNode if the first item
in its list of nodes is an AccessorNode. This matches how PostgreSQL
treats accessors and therefore leads to formatting, but is more logical
as well: it makes no sense for a node in an accessor to also be an
accessor list.

Add the `validateNode` recursive function to ast and use it to mak a
second pass over an AST before returning it from New. This also matches
the PostgreSQL behavior, where it checks for errors not caught by its
lexer or parser, notably "trailing junk" in numeric value and use of
array subscripts outside of an array index node.

Stop wrapping errors returned by ast, as they should be wrapped by the
parser, instead, and looked goofy starting with `parser: parser: xxx`
(or even `parser: ast: xxx`).

Add `setResult` to the lexer and use it to set the result from the
grammar. This allows the lexer to to additional error handling now that
`ast.New` can return an error from `validateNode`.

Add `NewUnaryOrNumber` to ast and have the grammar use it to create
unary plus and minus nodes, which can often be simplified to plain old
numeric nodes instead, again affecting formatting and following the
pattern in PostgreSQL.

Teach the lexer to return `scanner.EOF` when it encounters errors to as
to cut off further processing and immediately return. This prevents the
grammar from creating nodes that will cause problems, such as invalid
numbers, for which the AST constructors panic.

Fix an issue scanning identifiers where it consumed one too many
characters. This is another way in which using `text/scanner` without
relying on its identifier scanning ins a PITA.

Create separate errors for null byte errors, as the "invalid hex"
message was misleading: it's valid hex, but not supported by PostgreSQL.

Port the full suite of jsonpath I/O tests from PostgreSQL. Results are
identical except for the aforementioned trailing zeros in decimal
values, presence of `\v` in strings rather than its hex encoding, and
error messages.

Other minor changes:

*   Consistently use `$(GO)` in the `Makefile`.
*   Upgrade protoc-gen-go and therefore the generated protobuf.
*   Use `strconv.Quote` instead of `fmt.Sprintf` to quote strings.