Add path.Path with marshaling
Add the Path data type, currently a thin wrapper around ast.AST, but
with its interface defined. It includes a method to distinguish between
SQL-standard path expression and Postgres predicate check expressions,
as well as a method for the indexable Postgres operator to use with the
path.

It also includes implementations of useful marshaling interfaces:

*   sql.Scanner
*   sql.Valuer
*   encoding.TextMarshaler
*   encoding.TextUnmarshaler
*   encoding.BinaryMarshaler
*   encoding.BinaryUnmarshaler

Currently the text and binary marshaling are the same.

Meanwhile, ast provides the predicate check detection method used by
Path, and gains the `Root` method, which will be essential for rewriting
ASTs down the line.