Refactor and document all ast nodes
Document the full list of ast types that implement ast.Node, adding a
test that lists them and a code starter for recursively processing them
all.

Rename AccessorNode to AccessorList node, and change the method to get
its list of nodes from `Nodes()` to `Accessors()`.

Change the method to fetch the nodes from ArrayIndexNode from `Nodes()`
to `Subscripts()`.

Change the name of the UnaryNode's operand accessor method from `Node()`
to `Operand()`, and add the same accessor to RegexNode.