Improve numeric prefix error handling
In a [brief hackers discussion], Peter Eisentraut points out that the
0x, 0o, and 0b prefixes are integers only. This insight allows some
simplification of the handling of these prefixes: if no digits are found
after the prefix, it's invalid. If digits are found but then a `.`, halt
lexing, because it is not going to turn into a decimal number.

So short-circuit the processing appropriately, and switch to using
`digSep` instead of a temporary `srcPos` variable to detect whether
digits were scanned (I had overlooked it previously).

Update the relevant `TestScanNumbers` test cases, add a couple more, fix
some duplication of integer cases, and add a comment quoting the
Postgres output for all errors, so we can compare. In most cases the
error messages here are more informative, since our lexer goes into more
numeric detail than the Postgres lexer. Also quote all cases where the
lexing succeeds for only a portion of the path and add test cases to
`TestNumericEdgeCases` for the same paths to check that they parse the
same as Postgres.

Also: Remove the unused `'\''` case from `Lex()`, add a missing
`stopTok` return value on error, and add `IsLax()` and `IsStrict()` to
AST.

  [brief hackers discussion]: https://www.postgresql.org/message-id/flat/2948b6b2-aca9-4352-8697-0a76eb1594aa%40eisentraut.org#e3a4a64eb2151a3e4536f24c4f4e5148