Replace accessor list node with linked list
Follow the Postgres precedent to use a linked list of selectors rather
than a slice. This will simplify porting path execution, in which
individual nodes need to be aware if there is a next node to make the
proper recursive calls for a depth-first search.

All nodes are now structs that include a `next` field, and the interface
now includes the `setNext` and `Next` methods. The new `LinkNodes`
method replaces the `NewAccessorList` method, linking individual nodes
and return the head node.

Check for a next node in the `writeTo` method of each implementation of
Node and call its `writeTo` method if it exists. This allows the string
construction of a path to be properly generated by following the list of
nodes. Update a couple of places where `writeTo` also needs to be called
instead of `String` on other child nodes, since `String` does not call
`Next`.

Update the check for a predicate expression, although it is still not
right. Will be properly fixed in the next commit.