Correctly detect predicate queries at parsing time
The grammar borrowed from Postgres properly distinguishes between
predicates and expressions, so inject a shim there to set a boolean in
the lexer for predicate expressions. Then use it as a new argument to
`ast.New()` as a dependable, correct indication for a predicate.

Also update some old references to `strict`, which should be `lax`, as
that has been the parameter passed to `ast.New()` since e2d2f54