Add the types package with date and time types
In order to support the parsing and comparison of date, time, and
timestamp data types, add the `types` package. It uses a single
`parseTime` method for parsing of all date and time values, then
rejiggers the data as appropriate for each type. This matches the
behavior of PostgreSQL from the `executeDateTimeMethod` function in
`jsonpath_exec.c`. Logic for parsing different time offset formats
inspired by pgx/types.