Improve Postgres compatibility of DateTime types
Make a public interface, DateTime, with a required GoTime function that
returns a time.Time object. This will be used by path execution to group
the various date and time types and treat them uniformly where
appropriate.

Pass a precision argument to ParseTime and teach it to properly round to
that precision before returning a value. Rework that parsing to more
closely align with Postgres path execution parses different formats,
although the Go formats are a bit stricter, and create the appropriate
data type relative to the first format that succeeds. The executor will
need the different types, which correspond to Postgres data types.

Drop a comment with the start of a mapping from Postgres to Go date and
time parsing formats for future reference.

Change the return value from TimestampTZ.String to use the local time,
to correspond to how Postgres always displays timestamptz strings in the
configured time zone. This should allow one to set the TZ environment
variable to get the equivalent behavior.

Teach TimeTZ.Compare to more closely follow the Postgres pattern for
comparing timetz values, in which times are not considered equivalent if
they are in the same time zones. This contrasts with timestamptz, which
will allow them to be equivalent if they correspond go the same instant
in different zones. It's kind of horked, but the Postgres documentation
strongly warns people off the timetz type, since the SQL standard
poorly-defines its semantics. We follow suit.