Add missing methods to and fix Date method in AST
While working on path execution, I found some methods missing. This
commit adds them:

*   Add `Const()` method to ConstNode
*   Add `Name()` to MethodNode
*   Add `First()` and `Last()` to AnyNode

Also, replace MethodDate with UnaryDate. This allows all the date and
time methods to be of the same type (it was the only one that was a
MethodNode), which simplifies implementation of the executor. It does
require some tweaking of the grammar that varies a bit from the Postgres
original, but we already had some variation by separating methods from
Unary operators, so no big deal.

Other tweaks:

*   Fix capitalization of MethodBigInt
*   Update Postgres GitHub URLs to reference the `REL_17_BETA1` tag
*   Add printNode() to `parser_test.go`. I needed it for some debugging,
    but it's not ready for integration