Fix all possible linting issues in path/exec
Upgrade to golangci-lint v1.59.0 and update the config to eliminate new
warnings (mostly previously-deprecated linters that have now been
removed). Also disable nestif, since it's pretty redundant and more
aggressive than gocognit, tell errcheck to allow fmt functions without
error checks.

So rename `returnValue()` to `returnValueVerbose()`, which conditionally
returns a verbose error, and create a new `returnValue()` that checks
that an error is Verbose before conditionally returning it. Use it in
situations where the code doesn't know if an error is an `ErrVerbose`,
and use `returnValueVerbose()` when it does know it's an `ErrVerbose`.
Also change the string value of `ErrNull` to simply `NULL`, to echo how
`io.EOF` works.

Tweak the errors returned by exec to be more consistent, with separate
errors for execution, verbose execution (suppressible with
`WithSilent()`), and validity (for invalid AST or unexpected values
passed to functions). Audit all the errors to try to return them
consistently. Unfortunately, we can't remove the checking for verbosity
in the root functions, because `executeItemOptUnwrapResultSilent()`
(renamed from `executeItemOptUnwrapResultNoThrow()`) wants to disable
verbosity in certain very specific contexts.

Speaking of verbosity, rename `throwErrors` to `verbose`.

Work through all of the `//nolint` comments in the exec package and
refactor the code to allow them to be eliminated, or add a comment as to
why they're needed. This mostly involves moving various code blocks into
their own functions. This will simplify the forthcoming unit tests, as
well. Use a few other tricks to get conditionals and line counts down,
like adding `predFrom()` to convert a `bool` to a `predOutcome` in a
single line.

Rename `jVal` to `value`, mostly because it looks nicer, but also
because that's how the docs refer to them.