Split jsonpath execution into multiple files
The original single file was fine while porting from Postgres, but it's
nicer to split things up more or less according to categories of
execution. This will make it a bit easier to work with, and allow more
focused unit testing.

Document all of the functions and rename `ErrNull` to just `NULL`,
following the precedent of `io.EOF`.

While at it, tweak the `exhaustive` linter to consider switch statements
with a `default` to be complete and replace all suppression of
`goerr113` with its new name, `err113`.