Add time zone awareness to the datetime types
PostgreSQL relies on the value of the time zone GUC and defaults to the
cluster's system time when it's not set. Up to now we'd been relying on
the time zone behaviors of the time package and using the system local
time as the default time zone. This reduced compatibility, as there was
no way to customize the time zone like the Postgres GUC.

So add `ContextWithTZ` to allow a custom time zone to be added to the
context passed to the query functions, and add a context parameter to
the datetime parsing, converting, and comparing functions to read it.
This requires that the exec package pass the context objects around
quite a bit more, so do that, too. Oh, and fix the spelling of the file
named `conpare_test.go`.

The one place where passing the context was infeasible is the
TimestampTZ.String method, which must adhere to the `fmt.Stringer`
interface. So for TimestampTZ objects, store the time zone used when the
object is crated by NewTimestampTZ, and use it in the `String` method
(and only in the `String` method). This operates under the (reasonable)
assumption that the contextual time zone is unlikely to change between
the creation of a TimestampTZ object and its stringification.

Then `ToString(context.Context)` to the DateTime interface. This
function generates a string for the jsonpath `string()` method, which
for TimeTZ and TimestampTZ differs from `String()` in displaying only
the hours in an offset if there are no minutes. This corresponds to the
Postgres behavior, where the output of serialized JSON is otherwise
consistent with `String()`.

Next, remove the offset formats that include seconds, as Postgres
jsonpath never captures them or displays them, even though the Postgres
timetz and timestamptz types can store them. See the [pgsql-hackers
Discussion] for details.

Also update TimestampTZ and TimeTZ to store only offsets instead of full
time zone configurations. This aligns with the behavior of the Postgres
timestamptz and timetz types, which store only offsets. Full time zones
are used only when read from the context when parsing, stringifying, and
converting between datetime types.

Move the offset-adjusting functions into the new `offset.go` file along
with the context-handling functions, just to have the offset code all in
one place. Add explicit tests for them, too.

These changes allow all of the datetime incompatibilities in the tests
ported from PostgreSQL to be corrected, as they now simply add the time
zone to the context any time the Postgres tests set the GUC. They also
use the Postgres testing default time zone "PST8PDT" when it's not
explicitly set.

In the process, fix errors in casting and conversions between the
datetime types, move the casting logic into `To*` methods of each
datetime type, and adjust the tests to always use the datetime
constructors rather than create them directly.

Finally, remove the unused infinityModifier type, as Postgres jsonpath
does not support "inf" timestamps, and so was never used.

[pgsql-hackers Discussion]: https://www.postgresql.org/message-id/flat/A1454D02-F747-4A1E-8129-F7E940C7E4EB%40justatheory.com