Fully document path and fix DateTime issues
Revise the path README.md with proper working examples. Add
`internal/sqljson/path/test/readme_test.go` that tests each example, to
be sure they work. Added in a subdirectory so that the examples don't
themselves appear in the generated documentation; they're manually
maintained in the README. Make corrections where appropriate, such as
the escaping patterns required for regular expressions.

Add a Compatibility section to document all of the ways in which the
implementation varies from Postgres. It's quite detailed, and probably
missing some stuff, but I put in everything I could recall. Includes a
table comparing all the regular expression escapes in Postgres and the
Go regexp package and how they differ (quite a lot it turns out).

Speaking of regular expressions, fix their configuration so that all
those escapes work, especially character classes like `\d`, which, in
creating the examples, I found did not work. Fixed by expanding the
regexp configuration flags to include `syntax.ClassNL` and
`syntax.PerlX`.

Add `internal/sqljson/path/example_test.go` with detailed examples,
including documentation for the major methods, functions, and
configuration parameters. Include the PostgreSQL equivalents for clarity
of comparison.

Fix a few more DateTime issues discovered in this process, including
using a space instead of a "T" to separate date and time in timestamp
ToString output, as PostgreSQL does. Also fix what appears to be a bug
in the PostgreSQL implementation of casting from date and timestamp to
timestamptz (keeping an eye on [this thread] to see how Postgres
addresses it). This simplifies the casting a bit, deleting some fussy
code that tried to match the Postgres inconsistency.

Add detailed inline documentation to the path package itself, with
sections covering PostgreSQL equivalents, execution options, time zones,
the two types of queries, and errors. And given the potential confusion
about the two types of queries, add ExistsOrMatch, which dispatches to
the appropriate method depending on whether the query is a SQL standard
or predicate check expression.

Do the same for the types package, including examples comparing to
PostgreSQL behaviors, and fix a few additional issues discovered in the
process of writing them, notably that TimeTZ values compare to Time
values in the context of the current date (relative to the time zone
context).

  [this thread]: https://www.postgresql.org/message-id/flat/7DE080CE-6D8C-4794-9BD1-7D9699172FAB%40justatheory.com