Improve grammar coverage and fix keyvalue failure
TestExecVariable sometimes fails because the generated ID is non-
deterministic. So revamp the test to work around it by inspecting that
the generated ID is as expected.

Add tests to test heretofore un-covered `decimal()` method calls, and
replace a `panic()` with a proper parser error.

Add some tests for error string generation in `grammar.go`. Along with
the improved `decimal()` this gets coverage of grammar.y up to 91% (was
94% when testing `grammar.go` directly by adding the `-l` to the
`go:generate` line in `parser.go`).

While at it, remove irrelevant pre-commit configuration, as this
repository contains no Docker or shell script files.