Always adjust test keyvalue IDs at runtime
When IDs depend on calculating the delta between the memory address of a
container and a contained item, calculate them at runtime with the same
objects that the package uses to calculate them. This is necessary
because the memory layout can sometimes differ at runtime, which leads
to occasional test failures. The IDs should never differ once memory has
been allocated, so as long as we compare the same values it will work.

This requires creating the objects and/or arrays to compare in advance,
and determining their IDs in advance, so it gets to be a bit busy at the
start of those test functions, but works well and is the only sensible
approach.

This pattern is not required for IDs evenly divisible by `10000000000`,
or `0`, since they do not incorporate memory offsets.

There we a couple of existing instances of this pattern, so update them
to use the new `deltaBetween()` function to generate the IDs.