Add support for using the local time zone
Although parsing and loading the time zone is not implemented in the
time package available in JavaScript, `time.Local` is configured, which
is exactly what we want.

So rename the `TZ` option to `TZ Compare` and add the `Local TZ` option.
Teach the application to apply the local time zone when requested, and
remove all the existing TZ stuff, which did not work. Update the README
accordingly.