Reorder compats, update generators, rm unused CI
Reorder the Compatibilities list in path/README.md to be closer to the
order or priority: some items may be fixed in the future, and others
will not.

While at it, update the versions of `goyacc` and `stringer` to specific
versions, and remove some vestiges from `.golangci.yaml` of the app that
sqljson was forked from.