Make jsonpath .string() be immutable for datetimes
Just before the release of Postgres 17, a [discussion] of the effect of
DateStyle and TimeZone on datetime string formatting revealed that those
effects violated the immutability of the jsonpath functions. Commit
[cc4fdfa] fixed the issue by making the output of `.string()` on
datetimes the same as normal JSON output: always according to the time
zone with which a datetime was originally parsed, and always in ISO-8601
format.

Duplicate that change by removing the `ToString` method from the
DateTime interface and all of its implementations, replacing it with
`fmt.Stringer`, to always use `String`. Update the docs and tests in
line with [cc4fdfa], as well. Remove the paragraph about DateStyle from
the "Compatibility" section of the README, since it no longer impacts
datetime formatting.

While at it, add `.util/pglist.go`, a utility to generate a list of `git
diff` commands to run against the Postgres source when looking to sync
changes in the future. Use it to then replicate additional changes
between the previously-synced Postgres tag, `REL_17_BETA1` and the
recently-released `REL_17_2`, including:

*   A note about the behavior of lax mode on array evaluation in the
    README.
*   Update number links to the Postgres source.
*   Add a few links to `jsonpath_exec.c` and `func.sgml` for
    `.util/pglist.go` to find, as it was not previously linked, so would
    not be in the list of files to diff.
*   Use `boolean` instead of `bool` in error messages, as updated in
    postgres/postgres@f2353dd.
*   Move `pg2go.pl` to `.util`.
*   Replace the `String() string` method in the `ast.Node` interface
    with `fmt.Stringer`

  [discussion]: https://postgr.es/m/5e8879d0-a3c8-4be2-950f-d83aa2af953a@eisentraut.org
  [cc4fdfa]: https://github.com/postgres/postgres/commit/cc4fdfa
  [f2353dd]: https://github.com/postgres/postgres/commit/f2353dd